responsive and handle sending replies to the api

+++ assets/js/app.js
@@ -1,27 +1,36 @@
 const container = document.querySelector(".container")
 const userInfo = document.querySelector(".user_info")
 const commentForm = document.querySelector('.comment_form')
+const comment = document.querySelector('.comment_form textarea')
 const url = "http://localhost:3000/"
+let repliesArray = []
 
 //get the data from api
 const getUsers = async () => {
     const res = await fetch(`${url}comments`)
     const data = await res.json()
-    renderUsers(data)
+    return data
 }
 
 //set users info on UI
-const renderUsers = data => {
+const renderUsers = async () => {
+
+    //get users 
+    const data = await getUsers()
+
+    //render users on the UI
     data.forEach((item, index) => {
         let template = ""
+
         const { image, username } = item.user
+        const { score, createdAt, currentUser, id, content } = item
 
         template = `
         <div class="comment">
             <div class="comment_container">
                 <div class="vote">
                     <button class='vote_btn up_vote'><i class="far fa-thumbs-up"></i></button>  
-                    <p class="score">${item.score}</p>
+                    <p class="score">${score}</p>
                     <button class='vote_btn down_vote'><i class="far fa-thumbs-down"></i></button>  
                 </div>
 
@@ -30,45 +39,45 @@ const renderUsers = data => {
                         <div class="user_info">
                             <img class=avatar src=${image.webp} alt="${username} image">
                             <h3 class="name">${username}</h3>
-                            <p class="date">${item.createdAt}</p>
+                            <p class="date">${createdAt}</p>
                         </div>
                         
-                        ${item.currentUser ?
-                `<div class="header_buutos_container">
-                            <div onClick=edditComment() class="header_buttons">
+                        ${currentUser ?
+                `<div class="header_buutons_container">
+                            <div onClick=editComment(event) class="header_buttons edit_btn">
                                <img src="assets/images/icon-edit.svg" alt="edit icon">
                                <p>Edit</p>
                            </div>
-                           <div onClick=deleteComment(${item.id}) class="header_buttons">
+                           <div onClick=deleteComment(${id}) class="header_buttons delete-btn">
                               <img src="assets/images/icon-delete.svg" alt="delete icon">
                               <p>Delete</p>
                            </div>
                         </div>`  :
-                `<div onClick=handleReply() class="header_buttons" >
+                `<div onClick=replyForm(${id}) id=${id} class="header_buutons_container header_buttons reply_btn" >
                               <img src="assets/images/icon-reply.svg" alt="reply icon">
                               <p>Reply</p>
                         </div>`
-            } 
+            }
 
                     </div>
-                    <p class="comment_text">${item.content}</p>  
+                    <p class="comment_text">${content}</p>  
                 </div>
             </div>
+
            <div class="replies">
                 <div class="reply_line"></div>
-                <div class="replies_container">
-                    
-                </div>
+                <div class="replies_container"></div>
            </div>
         </div>
         `
         container.insertAdjacentHTML('beforeend', template)
 
-        if (item.replies.length) {
+        //check if there is any reply for the comment
+        if (item.replies.length > 0) {
             setReply(item.replies, index)
+            repliesArray.push(item.replies)
         }
     })
-
     voting()
 }
 
@@ -137,11 +146,11 @@ const voting = () => {
 //add comment
 const addComment = async (event) => {
     event.preventDefault()
-
+    if (commentForm.comment.value === '') return
     const doc = {
         currentUser: true,
         content: commentForm.comment.value,
-        createdAt: 2,
+        createdAt: new Date().toLocaleDateString(),
         score: Math.floor(Math.random() * 20),
         user: {
             image: {
@@ -160,16 +169,94 @@ const addComment = async (event) => {
     })
 }
 
-//eddit comment
-const edditComment = () => {
-    console.log(0);
+//edit comment
+const editComment = event => {
+
+    let editBtn = event.target
+
+    if (editBtn !== event.currentTarget) {
+        editBtn = event.target.parentElement
+    }
+    const textContainer = editBtn.parentNode.parentNode.nextElementSibling
+    let text = ''
+
+    if (textContainer.children.length > 0) {
+        text = textContainer.firstElementChild.innerHTML + ' ' + textContainer.lastElementChild.innerHTML
+    } else {
+        text = textContainer.innerHTML
+    }
+
+    template = `
+        <div class='edit_form_container'>
+      <form onSubmit=updateComment(event) class="edit_form">
+        <textarea name="comment">${text}</textarea>
+        <button>UPDATE</button>
+      </form>
+    </div>
+    `
+    textContainer.remove()
+    editBtn.parentNode.parentNode.parentNode.insertAdjacentHTML('beforeend', template)
+
+    editBtn.style.pointerEvents = 'none'
+
+    //check if there was a nother edit form open
+    document.querySelectorAll('.edit_form_container').forEach(form => {
+
+        if (editBtn.parentNode.parentNode.parentNode.lastElementChild !== form) {
+            const el = document.createElement('p')
+            el.classList.add('comment_text')
+            el.innerHTML = form.firstElementChild.firstElementChild.value
+            form.parentNode.appendChild(el)
+            form.remove()
+            document.querySelectorAll('.edit_btn').forEach(btn => btn.style.pointerEvents = 'all')
+            editBtn.style.pointerEvents = 'none'
+        }
+    })
+}
+
+//update comment
+const updateComment = (event) => {
+    event.preventDefault()
+
+    const updateValue = event.target.firstElementChild.value
+    const spliteValue = updateValue.split(' ')
+
+    if (updateValue === '') return
+
+    if (spliteValue[0].includes('@')) {
+
+        const container = document.createElement('div')
+
+        const text = document.createElement('p')
+        container.classList.add('comment_text')
+
+        const id = document.createElement('span')
+        id.classList.add('user_id')
+
+        text.innerHTML = spliteValue.slice(1, spliteValue.length).join(' ')
+        id.innerHTML = spliteValue[0]
+
+        container.append(id, text)
+        event.target.parentNode.parentNode.append(container)
+    }
+    else {
+        const text = document.createElement('p')
+        text.classList.add('comment_text')
+        text.innerHTML = updateValue
+        event.target.parentNode.parentNode.append(text)
+    }
+
+    event.target.parentNode.remove()
+    document.querySelectorAll('.edit_btn').forEach(btn => btn.style.pointerEvents = 'all')
 }
 
 //delete comment
-const deleteComment = async(id) => {
-    await fetch(`${url}comments/${id}`,{
-        method: 'DELETE'
-    })
+const deleteComment = async (id) => {
+    if (window.confirm('are you sure that you want to delete comment?')) {
+        await fetch(`${url}comments/${id}`, {
+            method: 'DELETE'
+        })
+    }
 }
 
 // show the replies on the UI
@@ -178,13 +265,15 @@ const setReply = (reply, index) => {
 
     reply.forEach(item => {
         let template = document.createElement('div')
+
         const { image, username } = item.user
+        const { score, createdAt, currentUser, id, content, replyingTo } = item
 
         template.innerHTML = `
-            <div class="reply_container">
+            <div class="reply_container" id='${id}'>
                 <div class="vote">
                     <button class='vote_btn up_vote'><i class="far fa-thumbs-up"></i></button>  
-                    <p class="score">${item.score}</p>
+                    <p class="score">${score}</p>
                     <button class='vote_btn down_vote'><i class="far fa-thumbs-down"></i></button>  
                 </div>
 
@@ -193,15 +282,26 @@ const setReply = (reply, index) => {
                             <div class="user_info">
                                 <img class=avatar src=${image.webp} alt="${username} image">
                                 <h3 class="name">${username}</h3>
-                                <p class="date">${item.createdAt}</p>
+                                <p class="date">${createdAt}</p>
                             </div>
 
-                            <div onClick=handleReply() class="header_buttons">
-                                <img src="assets/images/icon-reply.svg" alt="reply icon">
-                                <p>Reply</p>
-                            </div>    
+                               ${currentUser ?
+                `<div class="header_buutons_container">
+                            <div onClick=editComment(event) class="header_buttons edit_btn">
+                               <img src="assets/images/icon-edit.svg" alt="edit icon">
+                               <p>Edit</p>
+                           </div>
+                           <div onClick=deleteReply(${index},${id}) class="header_buttons delete-btn">
+                              <img src="assets/images/icon-delete.svg" alt="delete icon">
+                              <p>Delete</p>
+                           </div>
+                        </div>`  :''
+            }  
                     </div>
-                    <p class="reply_text"><span>@${item.replyingTo}</span> ${item.content}</p>
+                    <div class="reply_text">
+                    <span class='user_id'>@${replyingTo}</span>
+                    <p> ${content} </p>
+                    </div>
                 </div >
             </div >
     `
@@ -209,11 +309,138 @@ const setReply = (reply, index) => {
     })
 }
 
-// reply to oder comments
-const handleReply = () => {
-    console.log(0);
+// form for replying to other comments
+const replyForm = (id) => {
+
+    let template = ''
+    template = `
+        <div class="reply_form">
+      <form class="comment_form" onSubmit=handleReply(event,${id})>
+        <img class="avatar" src="assets/images/avatars/image-juliusomo.webp" alt="user avatar">
+        <textarea name="comment"></textarea>
+        <button>REPLY</button>
+      </form>
+    </div>
+    `
+    const replyBtn = document.querySelectorAll('.header_buttons')
+    replyBtn.forEach(btn => {
+        const container = btn.parentNode.parentNode.parentNode
+
+        if (parseInt(btn.id) === id) {
+            container.insertAdjacentHTML('afterend', template)
+            btn.style.pointerEvents = 'none'
+        } else if (container.nextElementSibling !== null) {
+            if (container.nextElementSibling.classList.contains('reply_form')) {
+                container.nextElementSibling.remove()
+                btn.style.pointerEvents = 'all'
+            }
+        }
+    })
 }
 
+//submit the reply
+const handleReply = async (event, id) => {
+    event.preventDefault()
+
+    //check if the form is open
+    const replyBtn = document.querySelectorAll('.header_buttons')
+    replyBtn.forEach(btn => {
+        btn.style.pointerEvents = 'all'
+    })
+
+    //check if the form is empty
+    if (event.target.children[1].value === '') {
+        document.querySelectorAll('.reply_form').forEach(form => {
+            form.remove()
+        })
+    } else {
+        //get users
+        const data = await getUsers()
+
+        data.forEach(async (user) => {
+            if (user.id == id) {
+
+                const doc = {
+                    replies: [
+                        ...user.replies,
+                        {
+                            id: Math.floor(Math.random() * Math.floor(Math.random() * Date.now())),
+                            currentUser: true,
+                            content: event.target.children[1].value,
+                            createdAt: new Date().toLocaleDateString(),
+                            score: Math.floor(Math.random() * 10),
+                            replyingTo: user.user.username,
+                            user: {
+                                image: {
+                                    png: "./assets/images/avatars/image-juliusomo.png",
+                                    webp: "./assets/images/avatars/image-juliusomo.webp"
+                                },
+                                username: "juliusomo"
+                            }
+                        }
+                    ]
+                }
+
+                await fetch(`${url}comments/${id}`, {
+                    method: 'PATCH',
+                    body: JSON.stringify(doc),
+                    headers: { 'Content-Type': 'application/json' }
+                })
+
+                repliesArray.push(doc)
+            }
+
+        })
+    }
+}
+
+//delete reply
+const deleteReply = async (index, id) => {
+    if (window.confirm('are you sure that you want to delete comment?')) {
+        if (repliesArray[index]) {
+            repliesArray[index] = repliesArray[index].filter(user => user.id != id)
+
+            await fetch(`${url}comments/${index + 1}`, {
+                method: 'PATCH',
+                body: JSON.stringify({ replies: [...repliesArray[index]] }),
+                headers: { 'Content-Type': 'application/json' }
+            })
+        }
+        else {
+            repliesArray[0] = repliesArray[0].filter(user => user.id != id)
+
+            await fetch(`${url}comments/${index + 1}`, {
+                method: 'PATCH',
+                body: JSON.stringify({ replies: [...repliesArray[0]] }),
+                headers: { 'Content-Type': 'application/json' }
+            })
+        }
+
+        if (repliesArray.length <= 1) {
+            await fetch(`${url}comments/${index + 1}`, {
+                method: 'PATCH',
+                body: JSON.stringify({ replies: [] }),
+                headers: { 'Content-Type': 'application/json' }
+            })
+        }
+    }
+}
+
+// check if there was a nother form open
+const clearForms = () => {
+    //delete reply forms
+    document.querySelectorAll('.comment_form').forEach(form => {
+        if (!form.classList.contains('main_form')) {
+            form.style.display = 'none'
+        }
+    })
+    document.querySelectorAll('.reply_btn').forEach(btn => btn.style.pointerEvents = 'all')
+}
+
+
 //events
-window.addEventListener("load", getUsers())
+window.addEventListener("load", renderUsers)
 commentForm.addEventListener('submit', addComment)
+comment.addEventListener('click', clearForms)
+
+
